chore(api): improve api reliability (#6)

+++ api/secret_santa_api/__init__.py
@@ -1,5 +1,7 @@
+import json
 import logging
 
+from dotenv import load_dotenv
 from environs import Env
 from flask import Flask, jsonify
 from flask_cors import CORS
@@ -11,13 +13,16 @@ from secret_santa_api.infrastructure.routes.draw import draw_bp
 from secret_santa_api.infrastructure.routes.participant import participant_bp
 
 
+load_dotenv(dotenv_path="./secret_santa_api/.env")
+
 env = Env()
 
 
-def create_app():
+def create_app(testing: bool = False):
     app = Flask(__name__)
+    app.url_map.strict_slashes = False
     app.errorhandler(Exception)(handle_error)
-    app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///secret_santa.db"
+    app.config["SQLALCHEMY_DATABASE_URI"] = env.str("SQLALCHEMY_DATABASE_URI")
     app.register_blueprint(draw_bp)
     app.register_blueprint(participant_bp)
 
@@ -26,8 +31,25 @@ def create_app():
     CORS(app, resources={r"*": {"origins": "http://localhost:5173"}})
 
     with app.app_context():
-        db.drop_all()
+        if testing:
+            db.drop_all()
         db.create_all()
+
+    @app.errorhandler(HTTPException)
+    def handle_werkzeug_http_exception(e):
+        """Return JSON instead of HTML for HTTP errors."""
+
+        response = e.get_response()
+        response.content_type = "application/json"
+        response.data = json.dumps(
+            {
+                "code": e.code,
+                "name": e.name,
+                "description": e.description,
+            }
+        )
+
+        return response
 
     return app
